Recurring tasks enhancements and small refactor
The logic behind the recurring tasks has been improved a lot. Also, now
editing recurring tasks is smart enough to ask to the user if he wants
to apply the changes to the single task or also the future related ones.

Also, DAO and DB have been improved simplifying the code, adding more
tests and covering almost very possible scenario to avoid issues with
edge cases.

Last but not least, during the code the improvable code was refactored.
I know that it's not ideal and it should be done apart, in another PR.
But since this is a side project sometimes you need to go dirty :)